+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## [10.9.1](https://github.com/nordnet/ui/compare/v10.9.0...v10.9.1) (2022-02-17)
+
+
+### Bug Fixes
+
+* 🐛 fix push error ([aaa76b0](https://github.com/nordnet/ui/commit/aaa76b0))
+* 🐛 remove 1px padding hack on mobile ([7739628](https://github.com/nordnet/ui/commit/7739628))
+* 🐛 remove quickfilter padding hack on mobile ([79ff0d1](https://github.com/nordnet/ui/commit/79ff0d1))
+
+# [10.9.0](https://github.com/nordnet/ui/compare/v10.8.1...v10.9.0) (2022-02-17)
+
+
+### Bug Fixes
+
+* also offset can be a function ([64db4b3](https://github.com/nordnet/ui/commit/64db4b3))
+* snapshots ([ade2ddb](https://github.com/nordnet/ui/commit/ade2ddb))
+* typo ([364e0f6](https://github.com/nordnet/ui/commit/364e0f6))
+
+
+### Features
+
+* pass offset to popover if supplied, update non ssr compliant utils method ([d4cd41f](https://github.com/nordnet/ui/commit/d4cd41f))
+
 ## [10.8.1](https://github.com/nordnet/ui/compare/v10.8.0...v10.8.1) (2022-02-16)